% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RastProcessing.R
\name{aggregate.rast}
\alias{aggregate.rast}
\title{Aggregate Rasters}
\usage{
\method{aggregate}{rast}()
}
\arguments{
\item{method}{character.}
}
\value{
SpatRaster-Stack
}
\description{
Aggregate spatial raster files for desired timespan (eg. daily, monthly, annual..)
}
