% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{aggregate_sensor}
\alias{aggregate_sensor}
\title{Aggregate Sensor}
\usage{
aggregate_sensor(
  data,
  sensors,
  station_ids,
  time_column,
  min_entries,
  coords = NULL
)
}
\arguments{
\item{data}{climate station data used to be aggregated}

\item{sensors}{character. Column names of your sensor columns.}

\item{station_ids}{character. Column name of your Climate Station IDs.}

\item{time_column}{character. The name of the time column in your climate station data.}

\item{min_entries}{minimum entries a time interval must have to be validly aggregated (e.g. 18 days for a monthly aggregation)}

\item{coords}{vector. If your climate station data already has coordinates, add coordinate column names in this argument (e.g. c("X", "Y"))}
}
\value{
data frame with aggregated climate station data
}
\description{
Description
}
\examples{
# example code

}
\seealso{
[prepClimateStations]
}
