% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RastProcessing.R
\name{calc.indices}
\alias{calc.indices}
\title{Calculate spectral indices}
\usage{
calc.indices(
  vi = "all",
  bands = c("blue", "green", "red", "nir", "nirb", "re1", "re2", "re3", "swir1", "swir2"),
  overwrite = FALSE
)
}
\arguments{
\item{vi}{Character. Either "all" or vector containing the preferred spectral indices. See 'Details' for more information.}

\item{bands}{Character. Vector with lenght(bands) = 10. Contains the names of the bands in the Raster Stack.
If bands from the *Usage* example vector dont exist, use "NA" in their position. See 'Details' for more information.}

\item{overwrite}{logical. Argument passed down from `terra`-package. Overwrite existing files?}
}
\value{
SpatRaster-Stack
}
\description{
Calculates a set of spectral indices to have more predictor variables available when further modeling.
}
\examples{
\dontrun{
calc.indices(vi = "all",
             bands = c("blue", "green", "red",
                       "nir", "nirb",
                       "re1", "re2", "re3",
                       "swir1", "swir2"),
             overwrite = TRUE)
}

}
\seealso{
`crop.all`, `fin.csv`
}
