% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{climplot}
\alias{climplot}
\title{Plotting}
\usage{
climplot(
  mnote,
  sensor,
  aoa = FALSE,
  mapcolors = rev(grDevices::terrain.colors(50)),
  scale_position = "bottomleft",
  north_position = "topright"
)
}
\description{
Plot results of climodr into maps
}
