% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{crop.all}
\alias{crop.all}
\title{Cropping all data}
\usage{
crop.all(
  folder_path = Input,
  crs = NULL,
  ext = NULL,
  method = "Smallest",
  safe_output = FALSE,
  ...
)
}
\arguments{
\item{folder_path}{charackter. Path to input folder. All necessary and relevant data should
be stored here. Default: Input created by [climodr::hubs].}

\item{crs}{Coordinate reference system Used to crop all images in folder_path. If crs
it will automatically reprojected into this one. Default: crs of smallest Extent.}

\item{ext}{SpatRaster, SpatVector or SpatExtent. Extent all data is cropped into. Default: Smallest Extent in folder_path.}

\item{method}{character.}

\item{safe_output}{logical. If cropped data should be safed permanently in the Environment put safe_output = TRUE.
Otherwise the output will be safed in the temporary directory. Default: FALSE.}
}
\value{
SpatRasters
}
\description{
Crops input data to the extent size
}
