% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictors.R
\name{elev.idx}
\alias{elev.idx}
\title{DEM-Based Predictors}
\usage{
elev.idx(proj = NULL, path = NULL)
}
\arguments{
\item{safe_output}{logical. If cleaned data should be safed permanently in the Environment put safe_output = TRUE.
Otherwise the output will be safed in the temporary directory. Default: FALSE.}
}
\value{
geotiff
}
\description{
Calculate DEM-Based Predictors
}
