% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_vignette.R
\docType{data}
\name{ext_vignette}
\alias{ext_vignette}
\title{Extent file for vignette}
\format{
## `ext_vignette`
\describe{
 \item{class}{SpatVector}
 \item{geometry}{polygons}
 \item{dimensions}{1, 1  (geometries, attributes)}
 \item{extent}{805737, 812824, 5890352, 5896005  (xmin, xmax, ymin, ymax)}
 \item{coord. ref.}{WGS 84 / UTM zone 32N (EPSG:32632)}
 \item{values}{1}
}
Spat Vector
}
\source{
Randomly created in (QGIS)[https://www.qgis.org/download/thank-you/].
}
\usage{
ext_vignette
}
\description{
A vector file containing the shape and extent of the Area used in the
vignette
}
\keyword{datasets}
