% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{extractPredictors}
\alias{extractPredictors}
\title{Extraction of Raster data at climate stations}
\usage{
extractPredictors(
  envrmt = .GlobalEnv$envrmt,
  climdata,
  rasters,
  unit,
  input = "Single",
  assign_raster_dates = "floor",
  save_output = FALSE
)
}
\arguments{
\item{envrmt}{variable name of your envrmt list created using climodr's \code{envi.create} function. Default = envrmt.}

\item{climdata}{Data frame. Climate station produced by spat.csv. Or produced via prepClimateStation. If null, climate statioon data is searched in Workflow/tworkflow.}

\item{rasters}{SpatRaster. Either a single raster with layers fitting for the desired time stamp. Or a bundled raster with fitting time stamps on the according layers. If empty, climodr searches for files created via \code{\link[=prepRasterData]{prepRasterData()}} in Output/rfinal.}

\item{unit}{character.}

\item{input}{character. Either "single" and "bundle". Depends on what output one has choose. "single" means a SpatRaster that only fits to one date, "bundle" is a SpatRaster contianing layers with time stamps fitting to the climate station data. Default = "bundle".}

\item{assign_raster_dates}{character. One of "floor", "ceiling" or "round". How should climate data be matched to a layer?
\itemize{
\item floor: Climate data is matched to the earliest date of the time interval (e.g. first day of a month)
\item ceiling: Climate data is matched to the latest date of the time interval (e.g. last day of a month)
\item round: Climate data is matched to the closest data available. (e.g. split between two months.)
}}

\item{...}{arguments passed down from other functions.}
}
\value{
data.frame
}
\description{
Extract the raster values of all raster layers from a scene at the station
coordinates at each time stamp. The extracted data will be attached to the
station data so there is a .csv-file with coordinates, sensor data (response values)
and extracted raster data (predictor values). The data is ready to be used for modelling.
}
\examples{
sdfsdf
}
\seealso{
\code{prep.csv}, \code{proc.csv}, \code{spat.csv}, \code{calc.indices}
}
