% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{final.csv}
\alias{final.csv}
\alias{fin.csv}
\title{Final aggregation for CSV-Data}
\usage{
fin.csv(method = "monthly", crs = NULL, safe_output = TRUE, ...)
}
\arguments{
\item{safe_output}{logical. If cleaned data should be safed permanently in the Environment put safe_output = TRUE.
Otherwise the output will be safed in the temporary directory. Default: FALSE.}
}
\value{
List
}
\description{
blabla
}
