% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{hubs}
\alias{hubs}
\title{Link input and output folders for climodr}
\usage{
hubs(in_dir, out_dir)
}
\arguments{
\item{in_dir}{character. Path to input folder. All necessary and relevant data should be stored here.}

\item{out_dir}{character. Path to output folder. Save all products in this destination.}
}
\description{
For easier and automated use of climodr. This function links either the
folder where all necessary input data is stored and also the folder,
where all relevant output data will be saved.
}
\examples{
\dontrun{
hubs("C:/Users/R_User/cool_proj/data/Input",
     "C:/Users/R_user/cool_proj/data/Output")
}

}
\seealso{
envi.create
}
