% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_description.R
\docType{data}
\name{plot_description}
\alias{plot_description}
\title{Plot description file}
\format{
## `plot_description`
A data frame with 10 rows and 5 columns
\describe{
 \item{plot}{imaginary station name and code}
 \item{general}{imaginary category of climate station}
 \item{region}{location name of climate station}
 \item{lat}{Latitude Coordinate of climate station}
 \item{lon}{Longitude Coordinate of climate station}
 \item{elevation}{elevation of climate station}
}
}
\source{
Randomly created coordinates and stations extracted from a random climate map.
}
\usage{
plot_description
}
\description{
Contains made up coordinates for imaginary climate stations for education purposes.
}
\keyword{datasets}
