% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RastProcessing.R
\name{prepRasterData}
\alias{prepRasterData}
\title{Prepare Raster Data}
\usage{
prepRasterData(
  envrmt = .GlobalEnv$envrmt,
  datepos,
  dateformat,
  crs = NULL,
  ext = NULL,
  mask = NULL,
  add_spectral_indices = FALSE,
  bands = NULL,
  output = "Single",
  overwrite = FALSE
)
}
\arguments{
\item{envrmt}{variable name of your envrmt list created using climodr's \link{envi.create} function. Default = envrmt.}

\item{datepos}{numeric vector. At which position in the filename does the date of your raster data start and where does it end?}

\item{dateformat}{character. The format of your date. See \code{\link[base:strptime]{base::strptime()}}}

\item{crs}{Coordinate reference system Used to crop all images in folder_path.}

\item{ext}{SpatRaster, SpatVector or SpatExtent. Extent all data is cropped into. Default: Smallest Extent in folder_path.}

\item{mask}{SpatVector or filename. Assign a polygon to mask your raster files. Either add file from global environment or the filename of your mask in Input/vector. Potentially reduces calculation times, as fewer pixels are contained in rasters after masking.}

\item{add_spectral_indices}{logical. Should spectral indices be calculated from input raster bands?}

\item{bands}{vector. If spectral indices should be calculated, this vector should contain the names of your required spectral bands. The order in the vector is fixed like this: c("blue", "green", "red", "nir", "swir"). Just switch out names in the vector with your equivalent band names.}

\item{output}{charakter. Either "single" or "bundle". "single" saves your dated raster data separately per date in Output/rfinal. "bundle" marks the corresponding layers with a time stamp and safes the dated rasters in a single file. Works more efficiently than "single", as long as there are more than one dates.}

\item{...}{arguments passed down from other functions.}
}
\value{
SpatRaster-Stack. Also saved to /workflow/rworkflow
}
\description{
Crops input data to the extent size and reprojects them into project Coordinate reference system.
}
\examples{
#create climodr environment and allow terra-functions to use 70\% of RAM


}
\seealso{
\code{fin.csv}, \code{calc.indices}
}
