% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{readClimateData}
\alias{readClimateData}
\title{Read climate station data}
\usage{
readClimateData(
  station_list,
  folder,
  station_ids,
  metadata = NULL,
  sep = ",",
  dec = "."
)
}
\arguments{
\item{station_list}{list. A list containing all paths of each climate station that should be read.}

\item{folder}{path. Folder where all the climate station data is stored in. Specifically made for climodr-environment. Either "$path_tabular" or "$path_vector".}

\item{station_ids}{character. The name of the station id column of your climate station data.}

\item{metadata}{vector. 5 Entries. Name of your metadata-file in your Input/dep folder, the column name for the climate stations, the name of your X and Y columns and the coordinate reference system. (e.g. c("metadata.csv", "plotID", "lon", "lat", "+proj=longlat +datum=WGS84")). NULL if data is vector_data.}

\item{sep}{character. Which seperator is used for the columns in the climate station data? (Default = ",")}

\item{dec}{character. Which symbol is used to mark decimal digits? (Default = ".")}
}
\value{
data frame or list. Depends if climate stations are vector files or tabular files. Tabular files return a single data frame, vector files return a list with the same data frame and a metadata vector, containting information about the metadata lost in conversion.
}
\description{
Readable climate station data for now: .csv, .txt, .gpkg, .shp
}
\examples{
# create climodr environment and allow terra-functions to use 70\% of RAM
envrmt <- envi.create(proj_path = tempdir(),
                      memfrac = 0.7)

# load example data
climsample(envrmt = envrmt)

# read climate station data
climate_stations <- list.files(envrmt$path_tabular, pattern = "Station")
csd <- readClimateData(station_list = climate_stations,
                       folder = envrmt$path_tabular,
                       station_ids = "plotID",
                       metadata = c("plot_description.csv",
                                    "plot",
                                    "x",
                                    "y",
                                    "+proj=longlat +datum=WGS84"))
head(csd)
}
\seealso{
[prepClimateStations]
}
