% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{splitTime}
\alias{splitTime}
\title{Split Time}
\usage{
splitTime(data, time_column, smallest_interval)
}
\arguments{
\item{data}{Data frame. Climate station data with time column in POSIXct-format}

\item{time_column}{Character. Column name of the column containing timestamps in POSIXct-format}

\item{smallest_interval}{The smallest time intervall the data should be split into. Same as units in [lubridate::floor_date()]}
}
\value{
data frame with aggregated climate station data
}
\description{
Description
}
\examples{
# example code

}
\seealso{
[prepClimateStations]
}
