% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station_W10.R
\docType{data}
\name{Station_W10}
\alias{Station_W10}
\title{Station File (W10)}
\format{
## `Station_W10`
A data frame with 10 rows and 5 columns
\describe{
 \item{plotID}{Plot ID of imaginary climate station}
 \item{datetime}{Timestamp of record for station}
 \item{Ta_200}{Imaginary air temperature at 200 cm above ground}
}
}
\source{
Randomly created climate data at random created stations extracted from a random climate map made by climodr.
}
\usage{
Station_W10
}
\description{
Contains made up climate data for imaginary climate stations for education purposes.
}
\keyword{datasets}
