% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autocorrelation.R
\name{autocorr}
\alias{autocorr}
\title{Autocorrelation}
\usage{
autocorr(method = "monthly", pred, resp, plot.corrplot = TRUE)
}
\arguments{
\item{eval_vec}{}
}
\description{
Test Data on autocorrelation to produce reliable models
}
