% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RastProcessing.R
\name{calc.indices}
\alias{calc.indices}
\title{Calculate spectral indices}
\usage{
calc.indices(
  vi = "all",
  bands = c("blue", "green", "red", "nir", "nirb", "re1", "re2", "re3", "swir1", "swir2"),
  overwrite = FALSE
)
}
\arguments{
\item{method}{character.}

\item{ext}{SpatRaster, SpatVector or SpatExtent. Extent all data is cropped into. Default: Smallest Extent in folder_path.}
}
\value{
SpatRaster-Stack
}
\description{
Calculates a set of spectral indices
}
