% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelling.R
\name{calc.model}
\alias{calc.model}
\title{Modelling}
\usage{
calc.model(
  method = "monthly",
  timespan,
  climresp,
  classifier = c("rf", "pls", "lm", "glm"),
  seed = NULL,
  p = 0.8,
  folds = "all",
  predrows,
  mnote = NULL,
  k = NULL,
  tc_method = "cv",
  metric = "RMSE",
  doParallel = FALSE,
  autocorrelation = FALSE,
  ...
)
}
\arguments{
\item{timespan}{vector or integer. Should contain all years to be modeled.
The years have to be the same format as in the tabular data.}

\item{climresp}{vector or integer. Should contain all response column's in
the tabular data.}

\item{classifier}{vector or character. Model variants to be used. Supported
models: Random Forest = "rf", Partial-Least-Squares = "pls",
Neural Networks = "nnet", Linear Regression = "lm" or
generalized boosted regression = "gbm".}

\item{seed}{integer. Seed to reproduce the same model over and over. e.g.
"seed = 707"}

\item{predrows}{vector or integer. Should contain the rows where all the
predictor values are stored in.}

\item{mnote}{character. Model note for special modifications used.
Default: "normal"}

\item{k}{integer. When 'fold' = "LLO" or "LTO". Set k to the number
of unique spatial or temporal units.}

\item{tc_method}{character. Method for train control function from caret
package. Default = "cv".}

\item{metric}{character. See "train".}

\item{doParallel}{logical. Parallelization accelerates the modelling
process. Warning: Your PC will slow down drastically. Make
sure to not run any other heavy processes during this.}

\item{fold}{character. Saving Algorithm. Either folding over location
only "LLO", over time only "LTO", or over both "LLTO".}

\item{safe_output}{logical. If cleaned data should be safed permanently in
the Environment put safe_output = TRUE. Otherwise the output
will be safed in the temporary directory. Default: FALSE.}
}
\value{
data frame.
}
\description{
Creates Models for each climate value
}
