% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RastProcessing.R
\name{crop.all}
\alias{crop.all}
\title{Cropping tiff data}
\usage{
crop.all(
  method = "MB_Timeseries",
  crs = NULL,
  ext = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{method}{charackter.}

\item{crs}{Coordinate reference system Used to crop all images in folder_path. If crs
it will automatically reprojected into this one. Default: crs of smallest Extent.}

\item{ext}{SpatRaster, SpatVector or SpatExtent. Extent all data is cropped into. Default: Smallest Extent in folder_path.}
}
\value{
SpatRaster-Stack
}
\description{
Crops input data to the extent size
}
