% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{envi.create}
\alias{envi.create}
\title{Create temporary environment}
\usage{
envi.create(proj_path)
}
\arguments{
\item{proj_path}{character. Path for working directory.}
}
\description{
Creates an environment climodr will use to store necessary, temporary data, which will be deleted after not being used anymore.
}
\examples{
\dontrun{
envi.create("C:/USERS/R_User/cool_proj")
}

}
\seealso{
[climodr::hubs] which climodr requires to be set up completely

[envimaR::alternativeEnvi], [envimaR::createEnvi] which this function wraps.
}
