% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{fin.csv}
\alias{fin.csv}
\title{Final aggregation for CSV-Data}
\usage{
fin.csv(method = "monthly", crs = NULL, save_output = TRUE, ...)
}
\arguments{
\item{method}{character. Either "daily", monthly" or "annual". Also depends on the available data.}

\item{save_output}{logical. If cleaned data should be saved permanently in the Environment put save_output = TRUE.
Otherwise the output will be saved in the temporary directory. Default: FALSE.}
}
\value{
List
}
\description{
Extract the raster values of all raster layers from a scene at the station
coordinates at each time stamp. The extracted data will be attached to the
station data so there is a .csv-file with coordinates, sensor data (response values)
and extracted raster data (predictor values). The data is ready to be used for modelling.
}
\examples{
\dontrun{
csv_fin <- fin.csv(method = "monthly",
                   save_output = TRUE)
head(csv_fin)
}

}
\seealso{
`prep.csv`, `proc.csv`, `spat.csv`, `calc.indices`
}
