% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{prep.csv}
\alias{prep.csv}
\title{Preparing CSV-Data}
\usage{
prep.csv(method = "proc", safe_output = TRUE, ...)
}
\arguments{
\item{method}{character. "proc" for ready-to-use data in separate .csv-files. "tube" for raw-data in the right format. "AI" for raw-data in other formats. Default "proc"-Method.}

\item{safe_output}{logical. If cleaned data should be safed permanently in the Environment put safe_output = TRUE.
Otherwise the output will be safed in the temporary directory. Default: FALSE.}
}
\value{
List
}
\description{
Crops input data to the extent size
}
