% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{prep.csv}
\alias{prep.csv}
\title{Preparing CSV-Data}
\usage{
prep.csv(method = "proc", save_output = TRUE, ...)
}
\arguments{
\item{method}{character. "proc" for ready-to-use data in separate .csv-files. "tube" for raw-data from the Tube Data Base. Default "proc"-Method.}

\item{save_output}{logical. If cleaned data should be saved permanently in the Environment put save_output = TRUE.
Otherwise the output will be saved in the temporary directory. Default: FALSE.}
}
\value{
List
}
\description{
Crops input data to the extent size and removes NA-Values
}
\examples{
\dontrun{
prep.csv(method = "proc", save_output = TRUE)

#check the created csv files
csv_files <- grep("_no_NAs.csv$",
                  list.files(envrmt$path_tworkflow),
                  value=TRUE)
csv_files
}

}
\seealso{
`proc.csv`, `spat.csv`, `fin.csv`
}
