% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{prepClimateStations}
\alias{prepClimateStations}
\title{Preparing Climate Station Data}
\usage{
prepClimateStations(
  envrmt = .GlobalEnv$envrmt,
  x,
  pattern = NULL,
  metadata = NULL,
  time_column,
  time_format,
  station_ids,
  sensors,
  aggregation_interval = NULL,
  start = NULL,
  end = NULL,
  crs = NULL,
  sep = ",",
  dec = ".",
  save_output = TRUE
)
}
\arguments{
\item{envrmt}{variable name of your envrmt list created using climodr's `envi.create` function. Default = envrmt.}

\item{x}{dataframe or climodr environment path. A data frame containing climate station data or one of the Input paths of the envrmt, where your climate station data is stored. Valid inputs for a climodr environment path are `"envrmt$path_tabular"` or `"envrmt$path_vector"`. The 'envrmt'-variable itself has to be the same as it is called in your global environment.}

\item{pattern}{character. Some string indicating that file is a climate station. E.g. "Climate_Station_*.csv" All climate station files in input folder must contain this pattern in filename.}

\item{metadata}{vector. 5 Entries. Name of your metadata-file in your Input/dep folder, the column name for the climate stations, the name of your X and Y columns and the coordinate reference system. (e.g. c("metadata.csv", "plotID", "lon", "lat", "+proj=longlat +datum=WGS84"))}

\item{time_column}{character. The name of the time column in your climate station data.}

\item{time_format}{character. How is your time column formated? (e.g. YYYY-MM-DD)}

\item{station_ids}{character. The name of the station id column of your climate station data.}

\item{sensors}{vector. Containing all column names of each sensor in your climate station data.}

\item{aggregation_interval}{character. To what format should your data be aggregated? Same as units in [lubridate::floor_date()]. Valid intervals are `second`, `minute`, `hour`, `day`, `week`, `month`, `bimonth`, `quarter`, `season`, `halfyear` and `year`.}

\item{start}{character. When do you want your climate station data to start? Has to be exactly the same format as your input climate station data.}

\item{end}{character. When do you want your climate station data to end? Has to be exactly the same format as your input cliamte station data.}

\item{crs}{character. Destined coordinate reference system. If crs = NULL, climod searches for "res_area.tif" in Input/dep.}

\item{sep}{character. Which seperator is used for the columns in the climate station data? (Default = ",")}

\item{dec}{character. Which symbol is used to mark decimal digits? (Default = ".")}

\item{...}{arguments passed down from other functions.}
}
\value{
data frame with aggregated climate station data
}
\description{
Crops input data to the extent size and removes NA-Values
}
\examples{
# create climodr environment and allow terra-functions to use 70\% of RAM
envrmt <- envi.create(proj_path = tempdir(),
                      memfrac = 0.7)

# load example data
climsample(envrmt = envrmt)

# prepare climate station data
climdata <- prepClimateStations(envrmt = envrmt,
                                x = envrmt$path_tabular,
                                pattern = "Station",
                                metadata = c("plot_description.csv",
                                             "plot",
                                             "lon",
                                             "lat",
                                             "+proj=longlat +datum=WGS84"),
                                time_column = "datetime",
                                time_format = "\%Y-\%m-\%dT\%H",
                                station_ids = "plotID",
                                sensors = "Ta_200",
                                aggregation_interval = "month")
head(climdata)
}
\seealso{
replacement function for the old [prep.csv], [proc.csv], [spat.csv] functions.
}
