% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{proc.csv}
\alias{proc.csv}
\title{Processing CSV-Data}
\usage{
proc.csv(method = "monthly", rbind = TRUE, safe_output = TRUE, ...)
}
\arguments{
\item{method}{character.}

\item{rbind}{logical. Create a single file with all climate stations. If FALSE, every station will be safed in a seperate file.}

\item{safe_output}{logical. If data should be safed permanently in the Environment put safe_output = TRUE.
Otherwise the output will be safed in the temporary directory. Default: FALSE.}
}
\value{
List
}
\description{
Calculate necessary Data from Stationary Data
}
